refactor: Cut out Request and Response object of listener.ts as request.ts and response.ts.
